+++ lib/api/utilities/api_retail_suppliers_offers.dart
@@ -58,6 +58,7 @@ class ApiRetailSuppliersOffers {
         'offerId',
         'region',
         'state',
+        'loadZone',
         'utility',
         'accountType',
         'countOfBillingCycles',