+++ test/db/utilities/retail_offers/retail_suppliers_offers_archive_test.dart
@@ -26,6 +26,7 @@ Future<void> tests(String rootUrl) async {
         'offerId',
         'region',
         'state',
+        'loadZone',
         'utility',
         'accountType',
         'countOfBillingCycles',